task

+++ app/Http/Controllers/TasksController.php
@@ -55,16 +55,18 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        // バリデーション
+       // バリデーション
         $request->validate([
+            'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
 
         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
-        $request->user()->tasks()->create([
-            'content' => $request->content,
-        ]);
-
+         // idの値で投稿を検索して取得
+          $request->user()->tasks()->create([
+                'status' => $request->status,
+                'content' => $request->content,
+            ]);
         // 前のURLへリダイレクトさせる
         return back();
     }
@@ -137,12 +139,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 }
